UML
I made a couple of abstract classes with Item, StatusEffects, and Character, which all serve as base clases for the core components for the game. The classes effectively organize and distribute the functionality of the code as a whole, and carry out the function of the player, monster, items, equipment, shop, and overall game structure. The cases that could cause the project to fail would be in a situation where a player tries to equip or use broken equipment, or the player tries to buy something when they do not have enough coins to do so. I have put in place a couple of measures to avoid any failures, by having some of the program do checks to see if the player can make an action that requires something like a certain coin amount, or if a weapon has enough durability. The overall UML relates to the project shows how each class connects to each other, and is a sort of rough structure of how the game will end up functioning. The UML demonstrates instances of inheritance, composition, and polymorphism and how that is being applied within my game classes.